Extract PaddingConverter to its own file
